Rebase of #54 upon versae/master
Fixed to work with newest version of Django.
Using StringIO to write one line at a time before flushing it is very slow. Reports download at about 5kbps. Also, previously the reports were created in memory and large reports would cause the server to run out of memory. This change causes the results to be written to the response while they are being read from the database. It never runs out of memory and runs at 600kbps.
The same change is on the HTML view as well. The same problem can present itself if the user selects to view all records. The raw cursor is passed instead of a list of results.